+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Disable SQLAlchemy logger
 - To avoid timeout errors, update genes, transcripts exons only from pre-downloaded files from schug
 - Documentation on how to update genes, transcripts and exons database tables
+- Renamed Pydantic `orm_mode` config param to `model_config`
 
 ## [3.1]
 ### Changed